Move pure algorithms to core module
