+++ settings.gradle.kts
@@ -3,5 +3,6 @@ rootProject.name = "Minecraft-Toolbox"
 include("buildSrc")
 findProject(":buildSrc")?.name = "gradle-plugin"
 
+include("core")
 include("base")
 include("plugin")
